+++ threedigrid_builder/interface/db.py
@@ -34,6 +34,8 @@ from threedigrid_builder.grid import Obstacles
 from threedigrid_builder.grid import Orifices
 from threedigrid_builder.grid import Pipes
 from threedigrid_builder.grid import Weirs
+from threedigrid_builder.grid import Surfaces
+from threedigrid_builder.grid import ImperviousSurfaces
 from typing import Callable
 from typing import ContextManager
 from typing import Tuple
@@ -249,6 +251,136 @@ class SQLite:
         target_epsg = self.epsg_code
         func = _get_reproject_func(SOURCE_EPSG, target_epsg)
         return pygeos.apply(geometries, func)
+
+    def get_surfaces(self) -> Surfaces:
+        with self.get_session() as session:
+            connection_node_map_array = (
+                session.query(models.ConnectionNode.id)
+                .order_by(models.ConnectionNode.id)
+                .as_structarray()
+            )
+
+            arr = (
+                session.query(
+                    models.Surface.id.label("surface_id"),
+                    models.Surface.function,
+                    models.Surface.code,
+                    models.Surface.display_name,
+                    models.Surface.nr_of_inhabitants,
+                    models.Surface.area,
+                    models.Surface.dry_weather_flow,
+                    models.Surface.the_geom,
+                    models.SurfaceParameter.outflow_delay,
+                    models.SurfaceParameter.surface_layer_thickness,
+                    models.SurfaceParameter.infiltration,
+                    models.SurfaceParameter.max_infiltration_capacity,
+                    models.SurfaceParameter.min_infiltration_capacity,
+                    models.SurfaceParameter.infiltration_decay_constant,
+                    models.SurfaceParameter.infiltration_recovery_constant,
+                    models.ConnectionNode.id.label("connection_node_id"),
+                    models.ConnectionNode.the_geom.label("connection_node_the_geom"),
+                    models.SurfaceMap.percentage,
+                )
+                .select_from(models.Surface)
+                .join(models.SurfaceParameter)
+                .join(
+                    models.SurfaceMap, models.SurfaceMap.surface_id == models.Surface.id
+                )
+                .join(
+                    models.ConnectionNode,
+                    models.SurfaceMap.connection_node_id == models.ConnectionNode.id,
+                )
+                .order_by(models.Surface.id)
+                .as_structarray()
+            )
+
+        # reproject
+        arr["the_geom"] = self.reproject(arr["the_geom"])
+        arr["connection_node_the_geom"] = self.reproject(
+            arr["connection_node_the_geom"]
+        )
+
+        # Map connection_node_id to row_id of connection node table
+        sort_idx = np.argsort(connection_node_map_array["id"])
+        connection_node_row_id = (
+            sort_idx[
+                np.searchsorted(
+                    connection_node_map_array["id"],
+                    arr["connection_node_id"],
+                    sorter=sort_idx,
+                )
+            ]
+        )
+
+        return Surfaces(
+            id=np.arange(0, len(arr["surface_id"] + 1), dtype=int),
+            connection_node_row_id=connection_node_row_id,
+            **{name: arr[name] for name in arr.dtype.names},
+        )
+
+    def get_impervious_surfaces(self) -> ImperviousSurfaces:
+        with self.get_session() as session:
+
+            connection_node_map_array = (
+                session.query(models.ConnectionNode.id)
+                .order_by(models.ConnectionNode.id)
+                .as_structarray()
+            )
+
+            arr = (
+                session.query(
+                    models.ImperviousSurface.id.label("surface_id"),
+                    models.ImperviousSurface.code,
+                    models.ImperviousSurface.display_name,
+                    models.ImperviousSurface.surface_inclination,
+                    models.ImperviousSurface.surface_class,
+                    models.ImperviousSurface.surface_sub_class,
+                    models.ImperviousSurface.nr_of_inhabitants,
+                    models.ImperviousSurface.area,
+                    models.ImperviousSurface.dry_weather_flow,
+                    models.ImperviousSurface.the_geom,
+                    models.ConnectionNode.id.label("connection_node_id"),
+                    models.ConnectionNode.the_geom.label("connection_node_the_geom"),
+                    models.ImperviousSurfaceMap.percentage,
+                )
+                .select_from(models.ImperviousSurface)
+                .join(models.ImperviousSurfaceMap)
+                .join(
+                    models.ConnectionNode,
+                    models.ImperviousSurfaceMap.connection_node_id
+                    == models.ConnectionNode.id,
+                )
+                .order_by(models.ImperviousSurface.id)
+                .as_structarray()
+            )
+
+        # convert enums to values
+        arr["surface_class"] = [x.value for x in arr["surface_class"]]
+        arr["surface_inclination"] = [x.value for x in arr["surface_inclination"]]
+
+        # reproject
+        arr["the_geom"] = self.reproject(arr["the_geom"])
+        arr["connection_node_the_geom"] = self.reproject(
+            arr["connection_node_the_geom"]
+        )
+
+        # Map connection_node_id to row_id of connection node table
+        sort_idx = np.argsort(connection_node_map_array["id"])
+        connection_node_row_id = (
+            sort_idx[
+                np.searchsorted(
+                    connection_node_map_array["id"],
+                    arr["connection_node_id"],
+                    sorter=sort_idx,
+                )
+            ]
+        )
+
+        return ImperviousSurfaces(
+            id=np.arange(0, len(arr["surface_id"] + 1), dtype=int),
+            connection_node_row_id=connection_node_row_id,
+            **{name: arr[name] for name in arr.dtype.names},
+        )
 
     def get_boundary_conditions_1d(self) -> BoundaryConditions1D:
         """Return BoundaryConditions1D"""