+++ threedigrid_builder/interface/gridadmin.py
@@ -11,6 +11,7 @@ from threedigrid_builder.constants import NodeType
 from threedigrid_builder.grid import Grid
 from threedigrid_builder.grid import GridMeta
 from threedigrid_builder.grid.cross_section_definitions import CrossSections
+from threedigrid_builder.base.surfaces import SurfaceMaps, Surfaces
 
 import numpy as np
 import pygeos
@@ -163,6 +164,8 @@ class GridAdminOut(OutputInterface):
         self.write_cross_sections(grid.cross_sections)
         self.write_levees(grid.levees)
         self.write_breaches(grid.breaches)
+        if grid.meta.has_0d:
+            self.write_surfaces(grid.surfaces, grid.surface_maps)
 
     def write_meta(self, meta: GridMeta):
         """Write the metadata to the gridadmin file.
@@ -490,6 +493,97 @@ class GridAdminOut(OutputInterface):
             group, "zoom_category", np.full(len(pumps), -9999, dtype="i4")
         )
 
+    def write_surfaces(self, surfaces: Surfaces, surface_maps: SurfaceMaps):
+
+        # For now use 0 instead of NaN as fill value for surfaces
+        # The calcore does not support NaN (for surfaces) yet
+        default_fill_value = 0
+
+        group = self._file.create_group("surface")
+        self.write_dataset(group, "id", surfaces.id, fill=default_fill_value)
+        self.write_dataset(group, "code", surfaces.code.astype("S100"), fill=b"")
+        self.write_dataset(
+            group, "display_name", surfaces.display_name.astype("S250"), fill=b""
+        )
+        self.write_dataset(group, "function", surfaces.function.astype("S64"), fill=b"")
+
+        self.write_dataset(
+            group,
+            "area",
+            surfaces.area,
+            fill=default_fill_value,
+            fill_nan=default_fill_value,
+        )
+        self.write_dataset(
+            group, "centroid_x", surfaces.centroid_x, fill=default_fill_value
+        )
+        self.write_dataset(
+            group, "centroid_y", surfaces.centroid_y, fill=default_fill_value
+        )
+        self.write_dataset(
+            group,
+            "dry_weather_flow",
+            surfaces.dry_weather_flow,
+            fill=default_fill_value,
+            fill_nan=default_fill_value,
+        )
+        self.write_dataset(
+            group,
+            "nr_of_inhabitants",
+            surfaces.nr_of_inhabitants,
+            fill=default_fill_value,
+            fill_nan=default_fill_value,
+        )
+        self.write_dataset(group, "infiltration_flag", surfaces.infiltration_flag)
+        self.write_dataset(
+            group,
+            "outflow_delay",
+            surfaces.outflow_delay,
+            fill=default_fill_value,
+            fill_nan=default_fill_value,
+        )
+        self.write_dataset(
+            group,
+            "storage_limit",
+            surfaces.storage_limit,
+            fill=default_fill_value,
+            fill_nan=default_fill_value,
+        )
+
+        if surfaces.surface_class is not None and np.any(surfaces.surface_class != None):  # noqa
+            # Impervious surfaces
+            self.write_dataset(
+                group, "surface_class", surfaces.surface_class.astype("S128"), fill=b""
+            )
+            self.write_dataset(
+                group,
+                "surface_inclination",
+                surfaces.surface_inclination.astype("S64"),
+                fill=b"",
+            )
+            self.write_dataset(
+                group,
+                "surface_sub_class",
+                surfaces.surface_sub_class.astype("S128"),
+                fill=b"",
+            )
+
+        # Surface params
+        self.write_dataset(group, "fb", surfaces.fb, fill=default_fill_value)
+        self.write_dataset(group, "fe", surfaces.fe, fill=default_fill_value)
+        self.write_dataset(group, "ka", surfaces.ka, fill=default_fill_value)
+        self.write_dataset(group, "kh", surfaces.kh, fill=default_fill_value)
+
+        # Surface maps (surface to connectionodes)
+        self.write_dataset(group, "fac", surface_maps.fac, fill=default_fill_value)
+        self.write_dataset(group, "nxc", surface_maps.nxc, fill=default_fill_value)
+        self.write_dataset(group, "nyc", surface_maps.nyc, fill=default_fill_value)
+        self.write_dataset(group, "pk", surface_maps.pk, fill=default_fill_value)
+        self.write_dataset(group, "cci", surface_maps.cci, fill=default_fill_value)
+        # Note: +1 for Fortran 1-based indexing
+        self.write_dataset(group, "cid", surface_maps.cid + 1, fill=default_fill_value)
+        self.write_dataset(group, "imp", surface_maps.imp + 1, fill=default_fill_value)
+
     def write_cross_sections(self, cross_sections: CrossSections):
         if len(cross_sections) == 0:
             return
@@ -534,7 +628,9 @@ class GridAdminOut(OutputInterface):
         self.write_dataset(group, "levmat", breaches.levmat)
         self.write_dataset(group, "coordinates", breaches.coordinates.T)
 
-    def write_dataset(self, group, name, values, fill=None, insert_dummy=True):
+    def write_dataset(
+        self, group, name, values, fill=None, insert_dummy=True, fill_nan=None
+    ):
         """Create the correct size dataset for writing to gridadmin.h5 and
         filling the extra indices with correct fillvalues.
 
@@ -543,7 +639,14 @@ class GridAdminOut(OutputInterface):
             name (str): Name of dataset
             values (array): Values of dataset to write.
             fill: fillvalue with same dtype as values.
+            fill_nane: (optional) fillvalue for nan values
         """
+
+        if fill_nan is not None:
+            if np.issubdtype(values.dtype, np.floating):
+                values = values.copy()
+                values[np.isnan(values)] = fill_nan
+
         if fill is None:
             if np.issubdtype(values.dtype, np.floating):
                 fill = np.nan